LayoutTests/http/wpt/wk-web-animations/timing-model/animation-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Check that the Animation currentTime property is set based on the timeline currentTime and the animation startTime</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

internals.pauseTimeline(document.timeline);

test(t => {
    const animation = new Animation;
    assert_equals(animation.currentTime, null);
}, "An animation's currentTime is null when the timeline has no currentTime set.");

test(t => {
    const animation = new Animation;
    animation.startTime = 2000;
    internals.setTimelineCurrentTime(document.timeline, 0);
    assert_equals(animation.currentTime, -2000);
    internals.setTimelineCurrentTime(document.timeline, 3000);
    assert_equals(animation.currentTime, 1000);
    animation.startTime = 1000;
    assert_equals(animation.currentTime, 2000);
}, "An animation accounts for its start time and document current time when computing its current time.");

test(t => {
    const animation = new Animation;
    animation.startTime = 1000;
    internals.setTimelineCurrentTime(document.timeline, 3000);
    assert_equals(animation.currentTime, 2000);

    animation.currentTime = 1000;
    assert_equals(animation.startTime, 2000);
}, "Setting the current time should adjust the start time.");

</script>
</body>
